Share more files between runtimes
