+++ src/native/clr/include/constants.hh
@@ -5,7 +5,7 @@
 
 #include <string_view>
 
-#include "shared/cpp-util.hh"
+#include <shared/cpp-util.hh>
 
 namespace xamarin::android {
 	class Constants
