+++ src/native/clr/include/runtime-base/logger.hh
@@ -5,7 +5,7 @@
 #include <string_view>
 
 #include <shared/log_types.hh>
-#include "strings.hh"
+#include <runtime-base/strings.hh>
 
 namespace xamarin::android {
 	class Logger
