+++ src/native/common/include/shared/helpers.hh
