+++ src/native/common/include/shared/xxhash.hh
