+++ src/native/mono/shared/cpp-util.hh
@@ -17,7 +17,7 @@
 #include <semaphore.h>
 #include <android/log.h>
 
-#include "helpers.hh"
+#include <shared/helpers.hh>
 
 namespace xamarin::android::detail {
 	[[gnu::always_inline, gnu::flatten]]
