+++ src/native/mono/shared/log_functions.cc
@@ -4,7 +4,7 @@
 #include <android/log.h>
 
 #include "java-interop-logger.h"
-#include "log_level.hh"
+#include <shared/log_level.hh>
 
 // Must match the same ordering as LogCategories
 static constexpr std::array<const char*, 12> log_names = {
