+++ src/native/mono/shared/log_types.hh
@@ -6,7 +6,7 @@
 #include <string_view>
 
 #include "java-interop-logger.h"
-#include "log_level.hh"
+#include <shared/log_level.hh>
 
 // We redeclare macros here
 #if defined(log_debug)
